refactored DayChangeListener
