+++ plep_main/src/deltadak/deletion/TaskDeletionInitialiser.kt
@@ -5,7 +5,7 @@ import deltadak.commands.DeleteCommand
 import deltadak.commands.DeleteSubtaskCommand
 import deltadak.commands.UndoFacility
 import deltadak.ui.Controller
-import deltadak.ui.util.STATIC.toHomeworkTaskList
+import deltadak.ui.util.toHomeworkTaskList
 import javafx.scene.control.ProgressIndicator
 import javafx.scene.control.TreeView
 import javafx.scene.input.KeyCode
