+++ plep_main/src/deltadak/ui/util/General.kt
@@ -1,6 +1,6 @@
 @file:Suppress("unused", "PackageName")
 
-package deltadak.ui.util.STATIC
+package deltadak.ui.util
 
 /**
  * Equals which returns true when both compared objects are null and is null safe.
