+++ plep_main/src/deltadak/ui/util/TaskRepeater.kt
@@ -1,4 +1,4 @@
-package deltadak.ui.util.STATIC
+package deltadak.ui.util
 
 import deltadak.Database
 import deltadak.HomeworkTask
